requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
pymarc==4.1.1
pysolr==3.9.0
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.6
-e git+https://github.com/herreio/vupysolr.git@215a0849512c06199fb3aacf87de97af213c8266#egg=vupysolr
